ApplyReactInfiniteToNoteTimeline
