+++ package.json
@@ -121,6 +121,7 @@
     "react-file-download": "^0.3.2",
     "react-fileupload-progress": "^0.4.0",
     "react-helmet": "^4.0.0",
+    "react-infinite": "^0.10.0",
     "react-lazyload": "^2.2.4",
     "react-motion": "^0.4.7",
     "react-onscroll": "^1.0.1",
@@ -140,9 +141,12 @@
     "redux-thunk": "^2.2.0",
     "rxjs": "^5.2.0",
     "slate": "^0.16.26",
+    "smooth-scroll": "github:cferdinandi/smooth-scroll",
+    "smoothy": "^1.1.2",
     "socket.io": "^1.7.3",
     "socket.io-client": "^1.7.3",
     "socket.io-react": "^1.1.2",
+    "sweet-scroll": "^2.2.0",
     "url": "^0.11.0",
     "webpack": "^2.2.1"
   },
