+++ src/actions/NoteTimelineActions.js
@@ -1,4 +1,4 @@
-import { UPDATE_TIMELINE_RENDER } from '../constants/index.js';
+import { UPDATE_TIMELINE_RENDER, CHANGE_TIMELINE_STATUS } from '../constants/index.js';
 
 
 
@@ -6,6 +6,14 @@ export function updateTimelineRender(value){
 
 	return { 
 		type: UPDATE_TIMELINE_RENDER,
+		data: value
+	}
+}
+
+export function changeTimelineUpdate(value){
+
+	return { 
+		type: CHANGE_TIMELINE_STATUS,
 		data: value
 	}
 }
