+++ src/constants/index.js
@@ -75,6 +75,7 @@ export const EDIT_FILE = "EDIT_FILE";
 export const NOTE_TIMELINE_GET = 'NOTE_TIMELINE_GET';
 export const NOTE_TIMELINE_SUCCESS = 'NOTE_TIMELINE_SUCCESS';
 export const NOTE_TIMELINE_CANCLE = 'NOTE_TIMELINE_CANCLE';
+export const CHANGE_TIMELINE_STATUS = 'CHANGE_TIMELINE_STATUS'
 export const NOTE_ONENOTE_GET = 'NOTE_ONENOTE_GET';
 export const NOTE_ONENOTE_SUCCESS = 'NOTE_ONENOTE_SUCCESS';
 export const NOTE_ONENOTE_CANCLE = 'NOTE_ONENOTE_CANCLE';
