+++ src/reducers/NoteTimelineReducer.js
@@ -1,7 +1,8 @@
 import { 
 	NOTE_TIMELINE_SUCCESS,
  	NOTE_ONENOTE_SUCCESS,
- 	UPDATE_TIMELINE_RENDER
+ 	UPDATE_TIMELINE_RENDER,
+ 	CHANGE_TIMELINE_STATUS
 } from '../constants/index.js'
 
 
@@ -11,6 +12,7 @@ var TimelineState = {
 	timelineRender: null,
 	status: "Loading",
 	HowManyNote: 0,
+	TimelineUpdate: false,
 	HowSorting: 'final_modified'
 	
 }
@@ -24,17 +26,21 @@ export function NoteTimeline(state = TimelineState, action) {
 			for (var key in action.data) {
 				timelineArray[Number(key)] = action.data[key]
 			}
+			
 			timelineArray.reverse()
 	
 			if (action.howSorting === 'final_modified') {
 
 				timelineArray.map( (a, index) => { a.timelineNum = index; return a } )
 			} 
+			// console.log('Soon Update timelnie', timelineArray)
+
 			return Object.assign({}, state, {
 				timeline: timelineArray,
 				status: "SUCCESS",
 				HowManyNote: timelineArray.length,
-				HowSorting: 'final_modified'
+				HowSorting: 'final_modified',
+				TimelineUpdate: true
 			})
 		case NOTE_ONENOTE_SUCCESS:
 			return Object.assign({}, state, {
@@ -50,6 +56,10 @@ export function NoteTimeline(state = TimelineState, action) {
 		case UPDATE_TIMELINE_RENDER:	
 			return Object.assign({}, state, {
 				timelineRender: action.data
+			})
+		case CHANGE_TIMELINE_STATUS:
+			return Object.assign({}, state, {
+				TimelineUpdate: action.data
 			})
 		default:
 			return state
